Refactor drawing
