+++ tessellation/draw.py
@@ -6,17 +6,11 @@ from typing import Union
 import matplotlib.pyplot as plt
 import numpy as np
 
+from tessellation.procgen import GenerationResult
+
 
 class Drawer:
     """Drawer class."""
-
-    def draw(self, *args, **kwargs):
-        """Draw the tessellation."""
-        raise NotImplementedError
-
-
-class MPLDrawer(Drawer):
-    """Matplotlib drawer class."""
 
     def __init__(self, cmap: str = "binary"):
         self.cmap = cmap
@@ -28,3 +22,33 @@ class MPLDrawer(Drawer):
     def save_as_png(self, file_path: Union[Path, str], tessellation: np.ndarray):
         """Save the drawing as a PNG file."""
         plt.imsave(file_path, tessellation, cmap=plt.get_cmap(self.cmap))
+
+    @staticmethod
+    def tessellate(
+        generation_result: GenerationResult, n_shapes: int = 5
+    ) -> np.ndarray:
+        """Tessellate the mask n_shapes times."""
+        mask = generation_result.mask
+        side_len = mask.shape[0]
+        side_len_full_image = side_len * n_shapes
+        tessellation = np.zeros((side_len_full_image, side_len_full_image), dtype=int)
+
+        color = 0
+        for i in range(n_shapes):
+            for j in range(n_shapes):
+                y_start = i * side_len
+                y_end = y_start + side_len
+
+                x_start = j * side_len
+                x_end = x_start + side_len
+
+                if color == 0:
+                    color_mask = np.logical_not(mask)
+                    color = 1
+                else:
+                    color_mask = mask
+                    color = 0
+
+                tessellation[y_start:y_end, x_start:x_end] = color_mask
+
+        return tessellation
